Moved two functions into utils. Will convert these to variables later.
