+++ knowledge_graph/make_graph.py
@@ -7,7 +7,7 @@ import pandas as pd
 from owlready2 import *
 
 
-from knowledge_graph.ontology_processing_utils import give_alias, save_test_ontology_to_json, save_graph_to_pickle
+from knowledge_graph.ontology_processing_utils import give_alias, save_test_ontology_to_json, save_graph_to_pickle, get_valid_test_ont, get_non_test_ont
 import os
 
 
@@ -129,7 +129,6 @@ def set_edge_properties(G):
                     to_remove[(node_b, prop)] = intersection
     return list(to_remove)
 
-
 def remove_edge_properties_from_nodes(G, to_remove):
     """Remove properties from Networkx nodes that occur on both nodes of an edge
     (because it marks that property is only for the edge).
@@ -148,6 +147,7 @@ def remove_edge_properties_from_nodes(G, to_remove):
             for node in list(G.nodes[node]["properties"][prop])
             if node not in list(to_delete)
         ]
+        # DM: uh... won't `node not in list(to_delete)` always evaluate to false?
 
 def remove_non_test_nodes(G, node, valid_test_ont, not_test_ont):
     if node in G.nodes:
@@ -169,55 +169,6 @@ def get_test_ontology(G, valid_test_ont, not_test_ont):
         node_b = edge[1]
         remove_non_test_nodes(G, node_a, valid_test_ont, not_test_ont)
         remove_non_test_nodes(G, node_b, valid_test_ont, not_test_ont)
-
-def get_valid_test_ont():
-    return {
-        "test ontology",
-        "personal value",
-        "achievement",
-        "benevolence",
-        "benevolence caring",
-        "benevolence dependability",
-        "conformity",
-        "conformity interpersonal",
-        "conformity rules",
-        "face",
-        "hedonism",
-        "humility",
-        "power",
-        "power dominance",
-        "power resources",
-        "security",
-        "security personal",
-        "security societal",
-        "self-direction",
-        "self-direction autonomy of action",
-        "self-direction autonomy of thought",
-        "stimulation",
-        "tradition",
-        "universalism",
-        "universalism concern",
-        "universalism nature",
-        "universalism tolerance",
-    }
-
-def get_non_test_ont():
-    return {
-        "value uncategorized (to do)",
-        "risk solution",
-        "adaptation",
-        "geoengineering",
-        "indirect adaptation",
-        "indirect geoengineering",
-        "indirect mitigration",
-        "carbon pricing",
-        "carbon tax",
-        "emissions trading",
-        "mitigation",
-        "solution to indirect adaptation barrier",
-        "solution to indirect mitigation barrier",
-        "solution uncategorized (to do)",
-    }
 
 def makeGraph(onto_path, edge_path, output_folder_path):
     """
